feat: on license request errors, return response body as cause (#137)
This adds a new cause property to the error object returned by player.error() when a license request fails with a 4xx or 5xx status code. It's a string representation, so, if it's JSON a consumer would need to call JSON.parse manually as we can't know whether it's JSON or not.