fix: convert initData to a string
This is to help get around the change in data type of initData between
the legacy API and the standards API.

In the legacy API, the initData was a Uint8Array but needed to be
converted to a Uint16Array to be able to toString it.

In the standards API, initData is an ArrayBuffer but can be converted to
a Uint8Array to be able to toString it.

We now do this string conversion for you so that you can just use the
contentId and not need to figure out whether legacy or standard API is
used.

BREAKING CHANGE: getContentId will now receive a string representation
of the initData